requirements.txt
mlflow==2.16.0
bcrypt==3.2.0
brotli==1.0.9
cloudpickle==2.2.1
configparser==5.2.0
cryptography==41.0.3
databricks-agents==0.5.0
databricks-feature-engineering==0.6.0
databricks-vectorsearch==0.40
google-cloud-storage==2.10.0
grpcio-status==1.60.0
gunicorn==20.1.0
langchain-community==0.2.4
langchain==0.2.1
numpy==1.23.5
pandas==1.5.3
psutil==5.9.0
pyarrow==14.0.1
pydantic==1.10.6
pyspark==3.5.0
rich==13.7.1
tornado==6.3.2